sql/descs: extract TableCollection into a new package
Ultimately I expect to call the TableCollection descs.Collection.

Release note: None
