+++ pkg/sql/internal.go
@@ -19,6 +19,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/kv"
 	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
+	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descs"
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgwirebase"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
@@ -68,7 +69,7 @@ type InternalExecutor struct {
 	// internal executor should see the modified schema. This interface allows
 	// the internal executor to modify its TableCollection to match the
 	// TableCollection of the parent executor.
-	tcModifier TableCollectionModifier
+	tcModifier descs.TableCollectionModifier
 }
 
 // MakeInternalExecutor creates an InternalExecutor.
