+++ pkg/sql/table_test.go
@@ -15,32 +15,11 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
-	"github.com/stretchr/testify/assert"
 )
-
-func TestIsSupportedSchemaName(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	testCases := []struct {
-		name  string
-		valid bool
-	}{
-		{"db_name", false},
-		{"public", true},
-		{"pg_temp", true},
-		{"pg_temp_1234_1", true},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.valid, isSupportedSchemaName(tree.Name(tc.name)))
-		})
-	}
-}
 
 func TestMakeTableDescColumns(t *testing.T) {
 	defer leaktest.AfterTest(t)()
